Bugfix/inconsistent locks (#99)
* BoundaryLock refactor

* first implementation

* conflicting locks issue fixes

* convert boundary lock fixes

* fix inconsistent locks issue + volume stress tests / short pointer issue

---------

+++ src/dbzero/core/collections/SGB_Tree/sgb_types.hpp
@@ -88,7 +88,7 @@ namespace db0
         typename ItemCompType, 
         typename ItemEqualType,
         typename CapacityType = std::uint16_t, 
-        typename AddressType = std::uint32_t, 
+        typename AddressType = std::uint64_t,
         typename HeaderType = db0::o_null,
         typename TreeHeaderType = db0::o_null>
     class sgb_types
